@@ -6,7 +6,7 @@ from collections.abc import Mapping
 from dataclasses import dataclass
 from functools import partial, singledispatch, wraps
 from types import MappingProxyType
-from typing import TYPE_CHECKING, Generic, TypeVar
+from typing import TYPE_CHECKING
 
 from anndata._io.utils import report_read_key_on_error, report_write_key_on_error
 from anndata._settings import settings
@@ -30,9 +30,7 @@ if TYPE_CHECKING:
 
     from ..._core.xarray import Dataset2D
 
-    T = TypeVar("T")
-    W = TypeVar("W", bound=_WriteInternal)
-    LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
+    type LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
 
 
 # TODO: This probably should be replaced by a hashable Mapping due to conversion b/w "_" and "-"
@@ -70,10 +68,10 @@ class IORegistryError(Exception):
         return cls(msg)
 
 
-def write_spec(spec: IOSpec):
+def write_spec[W: _WriteInternal](spec: IOSpec) -> Callable[[W], W]:
     def decorator(func: W) -> W:
         @wraps(func)
-        def wrapper(g: GroupStorageType, k: str, *args, **kwargs):
+        def wrapper(g: GroupStorageType, k: str, *args, **kwargs) -> None:
             result = func(g, k, *args, **kwargs)
             g[k].attrs.setdefault("encoding-type", spec.encoding_type)
             g[k].attrs.setdefault("encoding-version", spec.encoding_version)
@@ -84,20 +82,19 @@ def write_spec(spec: IOSpec):
     return decorator
 
 
-_R = TypeVar("_R", _ReadInternal, _ReadLazyInternal)
-R = TypeVar("R", Read, ReadLazy)
+class IORegistry[RI: (_ReadInternal, _ReadLazyInternal), R: (Read, ReadLazy)]:
+    read: dict[tuple[type, IOSpec, frozenset[str]], RI]
+    read_partial: dict[tuple[type, IOSpec, frozenset[str]], Callable]
+    write: dict[tuple[type, type | tuple[type, str], frozenset[str]], _WriteInternal]
+    write_specs: dict[type | tuple[type, str] | tuple[type, type], IOSpec]
 
+    def __init__(self) -> None:
+        self.read = {}
+        self.read_partial = {}
+        self.write = {}
+        self.write_specs = {}
 
-class IORegistry(Generic[_R, R]):
-    def __init__(self):
-        self.read: dict[tuple[type, IOSpec, frozenset[str]], _R] = {}
-        self.read_partial: dict[tuple[type, IOSpec, frozenset[str]], Callable] = {}
-        self.write: dict[
-            tuple[type, type | tuple[type, str], frozenset[str]], _WriteInternal
-        ] = {}
-        self.write_specs: dict[type | tuple[type, str] | tuple[type, type], IOSpec] = {}
-
-    def register_write(
+    def register_write[T](
         self,
         dest_type: type,
         src_type: type | tuple[type, str],
@@ -119,7 +116,7 @@ class IORegistry(Generic[_R, R]):
         else:
             self.write_specs[src_type] = spec
 
-        def _register(func):
+        def _register(func: _WriteInternal[T]) -> _WriteInternal[T]:
             self.write[(dest_type, src_type, modifiers)] = write_spec(spec)(func)
             return func
 
@@ -155,11 +152,11 @@ class IORegistry(Generic[_R, R]):
         src_type: type,
         spec: IOSpec | Mapping[str, str],
         modifiers: Iterable[str] = frozenset(),
-    ) -> Callable[[_R], _R]:
+    ) -> Callable[[RI], RI]:
         spec = proc_spec(spec)
         modifiers = frozenset(modifiers)
 
-        def _register(func):
+        def _register(func: RI) -> RI:
             self.read[(src_type, spec, modifiers)] = func
             return func
 
