@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from functools import cached_property
-from typing import TYPE_CHECKING, Generic, TypeVar
+from typing import TYPE_CHECKING
 
 import numpy as np
 import pandas as pd
@@ -30,11 +30,8 @@ if TYPE_CHECKING:
     from ...compat import Index1DNorm
 
 
-K = TypeVar("K", H5Array, ZarrArray)
-
-
-class ZarrOrHDF5Wrapper(XZarrArrayWrapper, Generic[K]):
-    def __init__(self, array: K):
+class ZarrOrHDF5Wrapper[K: (H5Array, ZarrArray)](XZarrArrayWrapper):
+    def __init__(self, array: K) -> None:
         self.chunks = array.chunks
         if isinstance(array, ZarrArray):
             super().__init__(array)
@@ -76,7 +73,7 @@ class ZarrOrHDF5Wrapper(XZarrArrayWrapper, Generic[K]):
         return self._array[key]
 
 
-class CategoricalArray(XBackendArray, Generic[K]):
+class CategoricalArray[K: (H5Array, ZarrArray)](XBackendArray):
     """
     A wrapper class meant to enable working with lazy categorical data.
     We do not guarantee the stability of this API beyond that guaranteed
@@ -131,7 +128,11 @@ class CategoricalArray(XBackendArray, Generic[K]):
         return pd.CategoricalDtype(categories=self.categories, ordered=self._ordered)
 
 
-class MaskedArray(XBackendArray, Generic[K]):
+# circumvent https://github.com/tox-dev/sphinx-autodoc-typehints/issues/580
+type K = H5Array | ZarrArray
+
+
+class MaskedArray[K: (H5Array, ZarrArray)](XBackendArray):
     """
     A wrapper class meant to enable working with lazy masked data.
     We do not guarantee the stability of this API beyond that guaranteed
