@@ -22,12 +22,10 @@ else:
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator, Sequence
-    from typing import TypeAlias, Union
 
     from scipy.sparse import spmatrix
 
-    # need to use Union because of autodoc_mock_imports
-    Array: TypeAlias = Union[torch.Tensor, np.ndarray, spmatrix]  # noqa: UP007
+    type Array = torch.Tensor | np.ndarray | spmatrix
 
 
 # Custom sampler to get proper batches instead of joined separate indices
