@@ -34,6 +34,7 @@ if TYPE_CHECKING:
     from typing import Any, Literal
 
     from .._core.merge import Reindexer, StrategiesLiteral
+    from .._types import Join_T
 
 SPARSE_MATRIX = {"csc_matrix", "csr_matrix"}
 
@@ -207,20 +208,25 @@ def write_concat_sparse(  # noqa: PLR0917
     axis: Literal[0, 1] = 0,
     reindexers: Reindexer | None = None,
     fill_value: Any = None,
-):
-    """
-    Writes and concatenates sparse datasets into a single output dataset.
+) -> None:
+    """Writes and concatenates sparse datasets into a single output dataset.
 
-    Args:
-        datasets (Sequence[BaseCompressedSparseDataset]): A sequence of BaseCompressedSparseDataset objects to be concatenated.
-        output_group (Union[ZarrGroup, H5Group]): The output group where the concatenated dataset will be written.
-        output_path (Union[ZarrGroup, H5Group]): The output path where the concatenated dataset will be written.
-        max_loaded_elems (int): The maximum number of sparse elements to load at once.
-        axis (Literal[0, 1], optional): The axis along which the datasets should be concatenated.
-            Defaults to 0.
-        reindexers (Reindexer, optional): A reindexer object that defines the reindexing operation to be applied.
-            Defaults to None.
-        fill_value (Any, optional): The fill value to use for missing elements. Defaults to None.
+    Parameters
+    ----------
+    datasets
+        A sequence of BaseCompressedSparseDataset objects to be concatenated.
+    output_group
+        The output group where the concatenated dataset will be written.
+    output_path
+        The output path where the concatenated dataset will be written.
+    max_loaded_elems
+        The maximum number of sparse elements to load at once.
+    axis
+        The axis along which the datasets should be concatenated.
+    reindexers
+        A reindexer object that defines the reindexing operation to be applied.
+    fill_value
+        The fill value to use for missing elements. Defaults to None.
     """
     elems = None
     if all(ri.no_change for ri in reindexers):
@@ -286,7 +292,7 @@ def _write_concat_arrays(  # noqa: PLR0913, PLR0917
     axis: Literal[0, 1] = 0,
     reindexers: list[Reindexer] | None = None,
     fill_value: Any = None,
-    join: Literal["inner", "outer"] = "inner",
+    join: Join_T = "inner",
 ):
     init_elem = arrays[0]
     init_type = type(init_elem)
@@ -328,10 +334,10 @@ def _write_concat_sequence(  # noqa: PLR0913, PLR0917
     output_path: str | Path,
     max_loaded_elems: int,
     axis: Literal[0, 1] = 0,
-    index: pd.Index = None,
+    index: pd.Index | None = None,
     reindexers: list[Reindexer] | None = None,
     fill_value: Any = None,
-    join: Literal["inner", "outer"] = "inner",
+    join: Join_T = "inner",
 ):
     """
     array, dataframe, csc_matrix, csc_matrix
@@ -411,7 +417,7 @@ def _write_axis_annot(  # noqa: PLR0917
     concat_indices: pd.Index,
     label: str,
     label_col: str,
-    join: Literal["inner", "outer"],
+    join: Join_T,
 ):
     concat_annot = pd.concat(
         unify_dtypes(read_elem(g[axis_name]) for g in groups),
@@ -447,7 +453,7 @@ def concat_on_disk(  # noqa: PLR0912, PLR0913, PLR0915
     *,
     max_loaded_elems: int = 100_000_000,
     axis: Literal["obs", 0, "var", 1] = 0,
-    join: Literal["inner", "outer"] = "inner",
+    join: Join_T = "inner",
     merge: StrategiesLiteral | Callable[[Collection[Mapping]], Mapping] | None = None,
     uns_merge: (
         StrategiesLiteral | Callable[[Collection[Mapping]], Mapping] | None
